Reorganize configtx unit tests
