+++ configtx/orderer_test.go
@@ -12,6 +12,7 @@ import (
 	"crypto/x509"
 	"encoding/base64"
 	"fmt"
+	"math/big"
 	"testing"
 
 	"github.com/golang/protobuf/proto"
@@ -1762,6 +1763,1368 @@ func TestGetOrdererOrg(t *testing.T) {
 
 	ordererOrgGroup := getOrdererOrg(config, "OrdererOrg")
 	gt.Expect(ordererOrgGroup).To(Equal(config.ChannelGroup.Groups[OrdererGroupKey].Groups["OrdererOrg"]))
+}
+
+func TestOrdererCapabilities(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				OrdererGroupKey: ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	ordererCapabilities, err := c.Orderer().Capabilities()
+	gt.Expect(err).NotTo(HaveOccurred())
+	gt.Expect(ordererCapabilities).To(Equal(baseOrdererConf.Capabilities))
+
+	// Delete the capabilities key and assert retrieval to return nil
+	delete(c.Orderer().ordererGroup.Values, CapabilitiesKey)
+	ordererCapabilities, err = c.Orderer().Capabilities()
+	gt.Expect(err).NotTo(HaveOccurred())
+	gt.Expect(ordererCapabilities).To(BeNil())
+}
+
+func TestAddOrdererCapability(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				OrdererGroupKey: ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	ordererOrgMSP := baseOrdererConf.Organizations[0].MSP
+	orgCertBase64, orgCRLBase64 := certCRLBase64(t, ordererOrgMSP)
+
+	expectedConfigGroupJSON := fmt.Sprintf(`{
+	"groups": {
+		"OrdererOrg": {
+			"groups": {},
+			"mod_policy": "Admins",
+			"policies": {
+				"Admins": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "MAJORITY",
+							"sub_policy": "Admins"
+						}
+					},
+					"version": "0"
+				},
+				"Endorsement": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "MAJORITY",
+							"sub_policy": "Endorsement"
+						}
+					},
+					"version": "0"
+				},
+				"Readers": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "ANY",
+							"sub_policy": "Readers"
+						}
+					},
+					"version": "0"
+				},
+				"Writers": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "ANY",
+							"sub_policy": "Writers"
+						}
+					},
+					"version": "0"
+				}
+			},
+			"values": {
+				"Endpoints": {
+					"mod_policy": "Admins",
+					"value": {
+						"addresses": [
+							"localhost:123"
+						]
+					},
+					"version": "0"
+				},
+				"MSP": {
+					"mod_policy": "Admins",
+					"value": {
+						"config": {
+							"admins": [
+								"%[1]s"
+							],
+							"crypto_config": {
+								"identity_identifier_hash_function": "SHA256",
+								"signature_hash_family": "SHA3"
+							},
+							"fabric_node_ous": {
+								"admin_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								},
+								"client_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								},
+								"enable": false,
+								"orderer_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								},
+								"peer_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								}
+							},
+							"intermediate_certs": [
+								"%[1]s"
+							],
+							"name": "MSPID",
+							"organizational_unit_identifiers": [
+								{
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								}
+							],
+							"revocation_list": [
+								"%[2]s"
+							],
+							"root_certs": [
+								"%[1]s"
+							],
+							"signing_identity": null,
+							"tls_intermediate_certs": [
+								"%[1]s"
+							],
+							"tls_root_certs": [
+								"%[1]s"
+							]
+						},
+						"type": 0
+					},
+					"version": "0"
+				}
+			},
+			"version": "0"
+		}
+	},
+	"mod_policy": "Admins",
+	"policies": {
+		"Admins": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "MAJORITY",
+					"sub_policy": "Admins"
+				}
+			},
+			"version": "0"
+		},
+		"BlockValidation": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "ANY",
+					"sub_policy": "Writers"
+				}
+			},
+			"version": "0"
+		},
+		"Readers": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "ANY",
+					"sub_policy": "Readers"
+				}
+			},
+			"version": "0"
+		},
+		"Writers": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "ANY",
+					"sub_policy": "Writers"
+				}
+			},
+			"version": "0"
+		}
+	},
+	"values": {
+		"BatchSize": {
+			"mod_policy": "Admins",
+			"value": {
+				"absolute_max_bytes": 100,
+				"max_message_count": 100,
+				"preferred_max_bytes": 100
+			},
+			"version": "0"
+		},
+		"BatchTimeout": {
+			"mod_policy": "Admins",
+			"value": {
+				"timeout": "0s"
+			},
+			"version": "0"
+		},
+		"Capabilities": {
+			"mod_policy": "Admins",
+			"value": {
+				"capabilities": {
+					"V1_3": {},
+					"V3_0": {}
+				}
+			},
+			"version": "0"
+		},
+		"ChannelRestrictions": {
+			"mod_policy": "Admins",
+			"value": null,
+			"version": "0"
+		},
+		"ConsensusType": {
+			"mod_policy": "Admins",
+			"value": {
+				"metadata": null,
+				"state": "STATE_NORMAL",
+				"type": "solo"
+			},
+			"version": "0"
+		}
+	},
+	"version": "0"
+}
+`, orgCertBase64, orgCRLBase64)
+
+	capability := "V3_0"
+	err = c.Orderer().AddCapability(capability)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	buf := bytes.Buffer{}
+	err = protolator.DeepMarshalJSON(&buf, &ordererext.DynamicOrdererGroup{ConfigGroup: c.Orderer().ordererGroup})
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	gt.Expect(buf.String()).To(Equal(expectedConfigGroupJSON))
+}
+
+func TestAddOrdererCapabilityFailures(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		testName     string
+		capability   string
+		ordererGroup func(og *cb.ConfigGroup)
+		expectedErr  string
+	}{
+		{
+			testName:   "when retrieving existing capabilities",
+			capability: "V1_3",
+			ordererGroup: func(og *cb.ConfigGroup) {
+				og.Values = map[string]*cb.ConfigValue{
+					CapabilitiesKey: {
+						Value: []byte("foobar"),
+					},
+				}
+			},
+			expectedErr: "retrieving orderer capabilities: unmarshaling capabilities: proto: can't skip unknown wire type 6",
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.testName, func(t *testing.T) {
+			t.Parallel()
+
+			gt := NewGomegaWithT(t)
+
+			baseOrdererConf, _ := baseSoloOrderer(t)
+			ordererGroup, err := newOrdererGroup(baseOrdererConf)
+			gt.Expect(err).NotTo(HaveOccurred())
+			tt.ordererGroup(ordererGroup)
+
+			config := &cb.Config{
+				ChannelGroup: &cb.ConfigGroup{
+					Groups: map[string]*cb.ConfigGroup{
+						OrdererGroupKey: ordererGroup,
+					},
+				},
+			}
+
+			c := New(config)
+
+			err = c.Orderer().AddCapability(tt.capability)
+			gt.Expect(err).To(MatchError(tt.expectedErr))
+		})
+	}
+}
+
+func TestRemoveOrdererCapability(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				OrdererGroupKey: ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	ordererOrgMSP := baseOrdererConf.Organizations[0].MSP
+	orgCertBase64, orgCRLBase64 := certCRLBase64(t, ordererOrgMSP)
+
+	expectedConfigGroupJSON := fmt.Sprintf(`{
+	"groups": {
+		"OrdererOrg": {
+			"groups": {},
+			"mod_policy": "Admins",
+			"policies": {
+				"Admins": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "MAJORITY",
+							"sub_policy": "Admins"
+						}
+					},
+					"version": "0"
+				},
+				"Endorsement": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "MAJORITY",
+							"sub_policy": "Endorsement"
+						}
+					},
+					"version": "0"
+				},
+				"Readers": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "ANY",
+							"sub_policy": "Readers"
+						}
+					},
+					"version": "0"
+				},
+				"Writers": {
+					"mod_policy": "Admins",
+					"policy": {
+						"type": 3,
+						"value": {
+							"rule": "ANY",
+							"sub_policy": "Writers"
+						}
+					},
+					"version": "0"
+				}
+			},
+			"values": {
+				"Endpoints": {
+					"mod_policy": "Admins",
+					"value": {
+						"addresses": [
+							"localhost:123"
+						]
+					},
+					"version": "0"
+				},
+				"MSP": {
+					"mod_policy": "Admins",
+					"value": {
+						"config": {
+							"admins": [
+								"%[1]s"
+							],
+							"crypto_config": {
+								"identity_identifier_hash_function": "SHA256",
+								"signature_hash_family": "SHA3"
+							},
+							"fabric_node_ous": {
+								"admin_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								},
+								"client_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								},
+								"enable": false,
+								"orderer_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								},
+								"peer_ou_identifier": {
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								}
+							},
+							"intermediate_certs": [
+								"%[1]s"
+							],
+							"name": "MSPID",
+							"organizational_unit_identifiers": [
+								{
+									"certificate": "%[1]s",
+									"organizational_unit_identifier": "OUID"
+								}
+							],
+							"revocation_list": [
+								"%[2]s"
+							],
+							"root_certs": [
+								"%[1]s"
+							],
+							"signing_identity": null,
+							"tls_intermediate_certs": [
+								"%[1]s"
+							],
+							"tls_root_certs": [
+								"%[1]s"
+							]
+						},
+						"type": 0
+					},
+					"version": "0"
+				}
+			},
+			"version": "0"
+		}
+	},
+	"mod_policy": "Admins",
+	"policies": {
+		"Admins": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "MAJORITY",
+					"sub_policy": "Admins"
+				}
+			},
+			"version": "0"
+		},
+		"BlockValidation": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "ANY",
+					"sub_policy": "Writers"
+				}
+			},
+			"version": "0"
+		},
+		"Readers": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "ANY",
+					"sub_policy": "Readers"
+				}
+			},
+			"version": "0"
+		},
+		"Writers": {
+			"mod_policy": "Admins",
+			"policy": {
+				"type": 3,
+				"value": {
+					"rule": "ANY",
+					"sub_policy": "Writers"
+				}
+			},
+			"version": "0"
+		}
+	},
+	"values": {
+		"BatchSize": {
+			"mod_policy": "Admins",
+			"value": {
+				"absolute_max_bytes": 100,
+				"max_message_count": 100,
+				"preferred_max_bytes": 100
+			},
+			"version": "0"
+		},
+		"BatchTimeout": {
+			"mod_policy": "Admins",
+			"value": {
+				"timeout": "0s"
+			},
+			"version": "0"
+		},
+		"Capabilities": {
+			"mod_policy": "Admins",
+			"value": {
+				"capabilities": {}
+			},
+			"version": "0"
+		},
+		"ChannelRestrictions": {
+			"mod_policy": "Admins",
+			"value": null,
+			"version": "0"
+		},
+		"ConsensusType": {
+			"mod_policy": "Admins",
+			"value": {
+				"metadata": null,
+				"state": "STATE_NORMAL",
+				"type": "solo"
+			},
+			"version": "0"
+		}
+	},
+	"version": "0"
+}
+`, orgCertBase64, orgCRLBase64)
+
+	capability := "V1_3"
+	err = c.Orderer().RemoveCapability(capability)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	buf := bytes.Buffer{}
+	err = protolator.DeepMarshalJSON(&buf, &ordererext.DynamicOrdererGroup{ConfigGroup: c.Orderer().ordererGroup})
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	gt.Expect(buf.String()).To(Equal(expectedConfigGroupJSON))
+}
+
+func TestRemoveOrdererCapabilityFailures(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		testName     string
+		capability   string
+		ordererGroup func(og *cb.ConfigGroup)
+		expectedErr  string
+	}{
+		{
+			testName:   "when capability does not exist",
+			capability: "V3_0",
+			ordererGroup: func(og *cb.ConfigGroup) {
+			},
+			expectedErr: "capability not set",
+		},
+		{
+			testName:   "when retrieving existing capabilities",
+			capability: "V3_0",
+			ordererGroup: func(og *cb.ConfigGroup) {
+				og.Values = map[string]*cb.ConfigValue{
+					CapabilitiesKey: {
+						Value: []byte("foobar"),
+					},
+				}
+			},
+			expectedErr: "retrieving orderer capabilities: unmarshaling capabilities: proto: can't skip unknown wire type 6",
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.testName, func(t *testing.T) {
+			t.Parallel()
+
+			gt := NewGomegaWithT(t)
+
+			baseOrdererConf, _ := baseSoloOrderer(t)
+			ordererGroup, err := newOrdererGroup(baseOrdererConf)
+			gt.Expect(err).NotTo(HaveOccurred())
+			tt.ordererGroup(ordererGroup)
+
+			config := &cb.Config{
+				ChannelGroup: &cb.ConfigGroup{
+					Groups: map[string]*cb.ConfigGroup{
+						OrdererGroupKey: ordererGroup,
+					},
+				},
+			}
+
+			c := New(config)
+
+			err = c.Orderer().RemoveCapability(tt.capability)
+			gt.Expect(err).To(MatchError(tt.expectedErr))
+		})
+	}
+}
+
+func TestOrdererOrg(t *testing.T) {
+	t.Parallel()
+	gt := NewGomegaWithT(t)
+
+	channel, _, _ := baseSystemChannelProfile(t)
+	channelGroup, err := newSystemChannelGroup(channel)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: channelGroup,
+	}
+
+	c := New(config)
+
+	expectedOrg := channel.Orderer.Organizations[0]
+
+	tests := []struct {
+		name        string
+		orgName     string
+		expectedErr string
+	}{
+		{
+			name:        "success",
+			orgName:     "OrdererOrg",
+			expectedErr: "",
+		},
+	}
+
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+			gt := NewGomegaWithT(t)
+
+			org, err := c.Orderer().Organization(tc.orgName).Configuration()
+			if tc.expectedErr != "" {
+				gt.Expect(err).To(MatchError(tc.expectedErr))
+				gt.Expect(Organization{}).To(Equal(org))
+			} else {
+				gt.Expect(err).NotTo(HaveOccurred())
+				gt.Expect(expectedOrg).To(Equal(org))
+			}
+		})
+	}
+}
+
+func TestRemoveOrdererOrg(t *testing.T) {
+	t.Parallel()
+	gt := NewGomegaWithT(t)
+
+	channel, _, _ := baseSystemChannelProfile(t)
+	channelGroup, err := newSystemChannelGroup(channel)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: channelGroup,
+	}
+
+	c := New(config)
+
+	c.Orderer().RemoveOrganization("OrdererOrg")
+	gt.Expect(c.Orderer().Organization("OrdererOrg")).To(BeNil())
+}
+
+func TestSetOrdererPolicy(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Orderer": ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	expectedPolicies := map[string]Policy{
+		ReadersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Readers",
+		},
+		WritersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Writers",
+		},
+		AdminsPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "MAJORITY Admins",
+		},
+		BlockValidationPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Writers",
+		},
+		"TestPolicy": {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Endorsement",
+		},
+	}
+
+	err = c.Orderer().SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "ANY Endorsement"})
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	updatedPolicies, err := c.Orderer().Policies()
+	gt.Expect(err).NotTo(HaveOccurred())
+	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
+}
+
+func TestSetOrdererPolicyFailures(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Orderer": ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	err = c.Orderer().SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{})
+	gt.Expect(err).To(MatchError("failed to set policy 'TestPolicy': unknown policy type: "))
+}
+
+func TestRemoveOrdererPolicy(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+	baseOrdererConf.Policies["TestPolicy"] = baseOrdererConf.Policies[AdminsPolicyKey]
+
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Orderer": ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	expectedPolicies := map[string]Policy{
+		ReadersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Readers",
+		},
+		WritersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Writers",
+		},
+		AdminsPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "MAJORITY Admins",
+		},
+		BlockValidationPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Writers",
+		},
+	}
+
+	err = c.Orderer().RemovePolicy("TestPolicy")
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	updatedPolicies, err := c.Orderer().Policies()
+	gt.Expect(err).NotTo(HaveOccurred())
+	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
+}
+
+func TestRemoveOrdererPolicyFailures(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+	baseOrdererConf.Policies["TestPolicy"] = baseOrdererConf.Policies[AdminsPolicyKey]
+
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				OrdererGroupKey: ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	tests := []struct {
+		testName      string
+		ordererGrpMod func(cb.ConfigGroup) *cb.ConfigGroup
+		policyName    string
+		expectedErr   string
+	}{
+		{
+			testName: "when removing blockvalidation policy",
+			ordererGrpMod: func(og cb.ConfigGroup) *cb.ConfigGroup {
+				return &og
+			},
+			policyName:  BlockValidationPolicyKey,
+			expectedErr: "BlockValidation policy must be defined",
+		},
+	}
+
+	for _, tt := range tests {
+		tt := tt
+		t.Run(tt.testName, func(t *testing.T) {
+			gt := NewGomegaWithT(t)
+
+			ordererGroup := tt.ordererGrpMod(*ordererGroup)
+			if ordererGroup == nil {
+				delete(config.ChannelGroup.Groups, OrdererGroupKey)
+			} else {
+				config.ChannelGroup.Groups[OrdererGroupKey] = ordererGroup
+			}
+
+			err = c.Orderer().RemovePolicy(tt.policyName)
+			gt.Expect(err).To(MatchError(tt.expectedErr))
+		})
+	}
+}
+
+func TestSetOrdererOrgPolicy(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Orderer": ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	expectedPolicies := map[string]Policy{
+		ReadersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Readers",
+		},
+		WritersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Writers",
+		},
+		AdminsPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "MAJORITY Admins",
+		},
+		EndorsementPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "MAJORITY Endorsement",
+		},
+		"TestPolicy": {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Endorsement",
+		},
+	}
+
+	ordererOrg := c.Orderer().Organization("OrdererOrg")
+	err = ordererOrg.SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "ANY Endorsement"})
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	updatedPolicies, err := ordererOrg.Policies()
+	gt.Expect(err).NotTo(HaveOccurred())
+	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
+}
+
+func TestSetOrdererOrgPolicyFailures(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Orderer": ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	err = c.Orderer().Organization("OrdererOrg").SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{})
+	gt.Expect(err).To(MatchError("unknown policy type: "))
+}
+
+func TestRemoveOrdererOrgPolicy(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	baseOrdererConf, _ := baseSoloOrderer(t)
+	baseOrdererConf.Organizations[0].Policies["TestPolicy"] = baseOrdererConf.Organizations[0].Policies[AdminsPolicyKey]
+
+	ordererGroup, err := newOrdererGroup(baseOrdererConf)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Orderer": ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	expectedPolicies := map[string]Policy{
+		ReadersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Readers",
+		},
+		WritersPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "ANY Writers",
+		},
+		AdminsPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "MAJORITY Admins",
+		},
+		EndorsementPolicyKey: {
+			Type: ImplicitMetaPolicyType,
+			Rule: "MAJORITY Endorsement",
+		},
+	}
+
+	err = c.Orderer().Organization("OrdererOrg").RemovePolicy("TestPolicy")
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	updatedPolicies, err := c.Orderer().Organization("OrdererOrg").Policies()
+	gt.Expect(err).NotTo(HaveOccurred())
+	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
+}
+
+func TestOrdererMSP(t *testing.T) {
+	t.Parallel()
+
+	gt := NewGomegaWithT(t)
+
+	soloOrderer, _ := baseSoloOrderer(t)
+	expectedMSP := soloOrderer.Organizations[0].MSP
+
+	ordererGroup, err := newOrdererGroup(soloOrderer)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				OrdererGroupKey: ordererGroup,
+			},
+		},
+	}
+
+	c := New(config)
+
+	msp, err := c.Orderer().Organization("OrdererOrg").MSP()
+	gt.Expect(err).NotTo(HaveOccurred())
+	gt.Expect(msp).To(Equal(expectedMSP))
+}
+
+func TestUpdateOrdererMSP(t *testing.T) {
+	t.Parallel()
+	gt := NewGomegaWithT(t)
+
+	channelGroup, privKeys, err := baseOrdererChannelGroup(t, orderer.ConsensusTypeSolo)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	config := &cb.Config{
+		ChannelGroup: channelGroup,
+	}
+	c := New(config)
+
+	ordererMSP, err := c.Orderer().Organization("OrdererOrg").MSP()
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	ordererCertBase64, ordererCRLBase64 := certCRLBase64(t, ordererMSP)
+
+	newRootCert, newRootPrivKey := generateCACertAndPrivateKey(t, "anotherca-org1.example.com")
+	newRootCertBase64 := base64.StdEncoding.EncodeToString(pemEncodeX509Certificate(newRootCert))
+	ordererMSP.RootCerts = append(ordererMSP.RootCerts, newRootCert)
+
+	newIntermediateCert, _ := generateIntermediateCACertAndPrivateKey(t, "anotherca-org1.example.com", newRootCert, newRootPrivKey)
+	newIntermediateCertBase64 := base64.StdEncoding.EncodeToString(pemEncodeX509Certificate(newIntermediateCert))
+	ordererMSP.IntermediateCerts = append(ordererMSP.IntermediateCerts, newIntermediateCert)
+
+	cert := ordererMSP.RootCerts[0]
+	certToRevoke, _ := generateCertAndPrivateKeyFromCACert(t, "org1.example.com", cert, privKeys[0])
+	signingIdentity := &SigningIdentity{
+		Certificate: cert,
+		PrivateKey:  privKeys[0],
+		MSPID:       "MSPID",
+	}
+	newCRL, err := c.Orderer().Organization("OrdererOrg").CreateMSPCRL(signingIdentity, certToRevoke)
+	gt.Expect(err).NotTo(HaveOccurred())
+	pemNewCRL, err := pemEncodeCRL(newCRL)
+	gt.Expect(err).NotTo(HaveOccurred())
+	newCRLBase64 := base64.StdEncoding.EncodeToString(pemNewCRL)
+	ordererMSP.RevocationList = append(ordererMSP.RevocationList, newCRL)
+
+	err = c.Orderer().Organization("OrdererOrg").SetMSP(ordererMSP)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	expectedConfigJSON := fmt.Sprintf(`
+{
+	"channel_group": {
+		"groups": {
+			"Orderer": {
+				"groups": {
+					"OrdererOrg": {
+						"groups": {},
+						"mod_policy": "Admins",
+						"policies": {
+							"Admins": {
+								"mod_policy": "Admins",
+								"policy": {
+									"type": 3,
+									"value": {
+										"rule": "MAJORITY",
+										"sub_policy": "Admins"
+									}
+								},
+								"version": "0"
+							},
+							"Endorsement": {
+								"mod_policy": "Admins",
+								"policy": {
+									"type": 3,
+									"value": {
+										"rule": "MAJORITY",
+										"sub_policy": "Endorsement"
+									}
+								},
+								"version": "0"
+							},
+							"Readers": {
+								"mod_policy": "Admins",
+								"policy": {
+									"type": 3,
+									"value": {
+										"rule": "ANY",
+										"sub_policy": "Readers"
+									}
+								},
+								"version": "0"
+							},
+							"Writers": {
+								"mod_policy": "Admins",
+								"policy": {
+									"type": 3,
+									"value": {
+										"rule": "ANY",
+										"sub_policy": "Writers"
+									}
+								},
+								"version": "0"
+							}
+						},
+						"values": {
+							"Endpoints": {
+								"mod_policy": "Admins",
+								"value": {
+									"addresses": [
+										"localhost:123"
+									]
+								},
+								"version": "0"
+							},
+							"MSP": {
+								"mod_policy": "Admins",
+								"value": {
+									"config": {
+										"admins": [
+											"%[1]s"
+										],
+										"crypto_config": {
+											"identity_identifier_hash_function": "SHA256",
+											"signature_hash_family": "SHA3"
+										},
+										"fabric_node_ous": {
+											"admin_ou_identifier": {
+												"certificate": "%[1]s",
+												"organizational_unit_identifier": "OUID"
+											},
+											"client_ou_identifier": {
+												"certificate": "%[1]s",
+												"organizational_unit_identifier": "OUID"
+											},
+											"enable": false,
+											"orderer_ou_identifier": {
+												"certificate": "%[1]s",
+												"organizational_unit_identifier": "OUID"
+											},
+											"peer_ou_identifier": {
+												"certificate": "%[1]s",
+												"organizational_unit_identifier": "OUID"
+											}
+										},
+										"intermediate_certs": [
+											"%[1]s",
+											"%[2]s"
+										],
+										"name": "MSPID",
+										"organizational_unit_identifiers": [
+											{
+												"certificate": "%[1]s",
+												"organizational_unit_identifier": "OUID"
+											}
+										],
+										"revocation_list": [
+											"%[3]s",
+											"%[4]s"
+										],
+										"root_certs": [
+											"%[1]s",
+											"%[5]s"
+										],
+										"signing_identity": null,
+										"tls_intermediate_certs": [
+											"%[1]s"
+										],
+										"tls_root_certs": [
+											"%[1]s"
+										]
+									},
+									"type": 0
+								},
+								"version": "0"
+							}
+						},
+						"version": "0"
+					}
+				},
+				"mod_policy": "Admins",
+				"policies": {
+					"Admins": {
+						"mod_policy": "Admins",
+						"policy": {
+							"type": 3,
+							"value": {
+								"rule": "MAJORITY",
+								"sub_policy": "Admins"
+							}
+						},
+						"version": "0"
+					},
+					"BlockValidation": {
+						"mod_policy": "Admins",
+						"policy": {
+							"type": 3,
+							"value": {
+								"rule": "ANY",
+								"sub_policy": "Writers"
+							}
+						},
+						"version": "0"
+					},
+					"Readers": {
+						"mod_policy": "Admins",
+						"policy": {
+							"type": 3,
+							"value": {
+								"rule": "ANY",
+								"sub_policy": "Readers"
+							}
+						},
+						"version": "0"
+					},
+					"Writers": {
+						"mod_policy": "Admins",
+						"policy": {
+							"type": 3,
+							"value": {
+								"rule": "ANY",
+								"sub_policy": "Writers"
+							}
+						},
+						"version": "0"
+					}
+				},
+				"values": {
+					"BatchSize": {
+						"mod_policy": "Admins",
+						"value": {
+							"absolute_max_bytes": 100,
+							"max_message_count": 100,
+							"preferred_max_bytes": 100
+						},
+						"version": "0"
+					},
+					"BatchTimeout": {
+						"mod_policy": "Admins",
+						"value": {
+							"timeout": "0s"
+						},
+						"version": "0"
+					},
+					"Capabilities": {
+						"mod_policy": "Admins",
+						"value": {
+							"capabilities": {
+								"V1_3": {}
+							}
+						},
+						"version": "0"
+					},
+					"ChannelRestrictions": {
+						"mod_policy": "Admins",
+						"value": null,
+						"version": "0"
+					},
+					"ConsensusType": {
+						"mod_policy": "Admins",
+						"value": {
+							"metadata": null,
+							"state": "STATE_NORMAL",
+							"type": "solo"
+						},
+						"version": "0"
+					}
+				},
+				"version": "0"
+			}
+		},
+		"mod_policy": "",
+		"policies": {},
+		"values": {},
+		"version": "0"
+	},
+	"sequence": "0"
+}`, ordererCertBase64, newIntermediateCertBase64, ordererCRLBase64, newCRLBase64, newRootCertBase64)
+
+	buf := bytes.Buffer{}
+	err = protolator.DeepMarshalJSON(&buf, c.updated)
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	gt.Expect(buf.String()).To(MatchJSON(expectedConfigJSON))
+}
+
+func TestUpdateOrdererMSPFailure(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		spec        string
+		mspMod      func(MSP) MSP
+		orgName     string
+		expectedErr string
+	}{
+		{
+			spec: "updating msp name",
+			mspMod: func(msp MSP) MSP {
+				msp.Name = "thiscantbegood"
+				return msp
+			},
+			orgName:     "OrdererOrg",
+			expectedErr: "MSP name cannot be changed",
+		},
+		{
+			spec: "invalid root ca cert keyusage",
+			mspMod: func(msp MSP) MSP {
+				msp.RootCerts = []*x509.Certificate{
+					{
+						SerialNumber: big.NewInt(7),
+						KeyUsage:     x509.KeyUsageKeyAgreement,
+					},
+				}
+				return msp
+			},
+			orgName:     "OrdererOrg",
+			expectedErr: "invalid root cert: KeyUsage must be x509.KeyUsageCertSign. serial number: 7",
+		},
+	}
+
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.spec, func(t *testing.T) {
+			t.Parallel()
+			gt := NewGomegaWithT(t)
+
+			channelGroup, _, err := baseOrdererChannelGroup(t, orderer.ConsensusTypeSolo)
+			gt.Expect(err).NotTo(HaveOccurred())
+
+			config := &cb.Config{
+				ChannelGroup: channelGroup,
+			}
+			c := New(config)
+
+			ordererMSP, err := c.Orderer().Organization("OrdererOrg").MSP()
+			gt.Expect(err).NotTo(HaveOccurred())
+
+			ordererMSP = tc.mspMod(ordererMSP)
+			err = c.Orderer().Organization(tc.orgName).SetMSP(ordererMSP)
+			gt.Expect(err).To(MatchError(tc.expectedErr))
+		})
+	}
 }
 
 func baseOrdererOfType(t *testing.T, ordererType string) (Orderer, []*ecdsa.PrivateKey) {
