+++ configtx/organization_test.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/hyperledger/fabric-config/protolator"
 	"github.com/hyperledger/fabric-config/protolator/protoext/ordererext"
-	cb "github.com/hyperledger/fabric-protos-go/common"
 	. "github.com/onsi/gomega"
 )
 
@@ -29,211 +28,6 @@ func TestOrganization(t *testing.T) {
 	org, err := getOrganization(orgGroup, "Org1")
 	gt.Expect(err).NotTo(HaveOccurred())
 	gt.Expect(expectedOrg).To(Equal(org))
-}
-
-func TestApplicationOrg(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel := Channel{
-		Consortium: "SampleConsortium",
-		Application: Application{
-			Policies:      standardPolicies(),
-			Organizations: []Organization{baseApplicationOrg(t)},
-		},
-	}
-	channelGroup, err := newChannelGroup(channel)
-	gt.Expect(err).NotTo(HaveOccurred())
-	orgGroup, err := newApplicationOrgConfigGroup(channel.Application.Organizations[0])
-	gt.Expect(err).NotTo(HaveOccurred())
-	channelGroup.Groups[ApplicationGroupKey].Groups["Org1"] = orgGroup
-
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	expectedOrg := channel.Application.Organizations[0]
-
-	tests := []struct {
-		name    string
-		orgName string
-	}{
-		{
-			name:    "success",
-			orgName: "Org1",
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			gt := NewGomegaWithT(t)
-
-			org, err := c.Application().Organization(tc.orgName).Configuration()
-			gt.Expect(err).ToNot(HaveOccurred())
-			gt.Expect(expectedOrg).To(Equal(org))
-		})
-	}
-}
-
-func TestRemoveApplicationOrg(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel := Channel{
-		Consortium: "SampleConsortium",
-		Application: Application{
-			Policies:      standardPolicies(),
-			Organizations: []Organization{baseApplicationOrg(t)},
-		},
-	}
-	channelGroup, err := newChannelGroup(channel)
-	gt.Expect(err).NotTo(HaveOccurred())
-	orgGroup, err := newOrgConfigGroup(channel.Application.Organizations[0])
-	gt.Expect(err).NotTo(HaveOccurred())
-	channelGroup.Groups[ApplicationGroupKey].Groups["Org1"] = orgGroup
-
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	c.Application().RemoveOrganization("Org1")
-	gt.Expect(c.updated.ChannelGroup.Groups[ApplicationGroupKey].Groups["Org1"]).To(BeNil())
-}
-
-func TestOrdererOrg(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel, _, _ := baseSystemChannelProfile(t)
-	channelGroup, err := newSystemChannelGroup(channel)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	expectedOrg := channel.Orderer.Organizations[0]
-
-	tests := []struct {
-		name        string
-		orgName     string
-		expectedErr string
-	}{
-		{
-			name:        "success",
-			orgName:     "OrdererOrg",
-			expectedErr: "",
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			gt := NewGomegaWithT(t)
-
-			org, err := c.Orderer().Organization(tc.orgName).Configuration()
-			if tc.expectedErr != "" {
-				gt.Expect(err).To(MatchError(tc.expectedErr))
-				gt.Expect(Organization{}).To(Equal(org))
-			} else {
-				gt.Expect(err).NotTo(HaveOccurred())
-				gt.Expect(expectedOrg).To(Equal(org))
-			}
-		})
-	}
-}
-
-func TestRemoveOrdererOrg(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel, _, _ := baseSystemChannelProfile(t)
-	channelGroup, err := newSystemChannelGroup(channel)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	c.Orderer().RemoveOrganization("OrdererOrg")
-	gt.Expect(c.Orderer().Organization("OrdererOrg")).To(BeNil())
-}
-
-func TestConsortiumOrg(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel, _, _ := baseSystemChannelProfile(t)
-	channelGroup, err := newSystemChannelGroup(channel)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	expectedOrg := channel.Consortiums[0].Organizations[0]
-
-	tests := []struct {
-		name           string
-		consortiumName string
-		orgName        string
-		expectedErr    string
-	}{
-		{
-			name:           "success",
-			consortiumName: "Consortium1",
-			orgName:        "Org1",
-			expectedErr:    "",
-		},
-	}
-
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-			gt := NewGomegaWithT(t)
-
-			org, err := c.Consortium(tc.consortiumName).Organization(tc.orgName).Configuration()
-			if tc.expectedErr != "" {
-				gt.Expect(Organization{}).To(Equal(org))
-				gt.Expect(err).To(MatchError(tc.expectedErr))
-			} else {
-				gt.Expect(err).ToNot(HaveOccurred())
-				gt.Expect(expectedOrg).To(Equal(org))
-			}
-		})
-	}
-}
-
-func TestRemoveConsortiumOrg(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel, _, _ := baseSystemChannelProfile(t)
-	channelGroup, err := newSystemChannelGroup(channel)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	c.Consortium("Consortium1").RemoveOrganization("Org1")
-	gt.Expect(c.Consortium("Consortium1").Organization("Org1")).To(BeNil())
 }
 
 func TestNewOrgConfigGroup(t *testing.T) {
