+++ configtx/policies_test.go
@@ -49,723 +49,6 @@ func TestPolicies(t *testing.T) {
 	gt.Expect(map[string]Policy{}).To(Equal(policies))
 }
 
-func TestRemoveApplicationOrgPolicy(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	applicationGroup := newConfigGroup()
-
-	application, _ := baseApplication(t)
-
-	for _, org := range application.Organizations {
-		org.Policies = applicationOrgStandardPolicies()
-		org.Policies["TestPolicy"] = Policy{
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		}
-
-		orgGroup, err := newOrgConfigGroup(org)
-		gt.Expect(err).NotTo(HaveOccurred())
-
-		applicationGroup.Groups[org.Name] = orgGroup
-	}
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	application.Organizations[0].Policies = applicationOrgStandardPolicies()
-	expectedOrgConfigGroup, _ := newOrgConfigGroup(application.Organizations[0])
-	expectedPolicies := expectedOrgConfigGroup.Policies
-
-	applicationOrg1 := c.Application().Organization("Org1")
-	err := applicationOrg1.RemovePolicy("TestPolicy")
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	actualOrg1Policies := applicationOrg1.orgGroup.Policies
-	gt.Expect(actualOrg1Policies).To(Equal(expectedPolicies))
-}
-
-func TestRemoveApplicationOrgPolicyFailures(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	applicationGroup := newConfigGroup()
-
-	application, _ := baseApplication(t)
-	for _, org := range application.Organizations {
-		org.Policies = applicationOrgStandardPolicies()
-		orgGroup, err := newOrgConfigGroup(org)
-		gt.Expect(err).NotTo(HaveOccurred())
-		applicationGroup.Groups[org.Name] = orgGroup
-	}
-
-	applicationGroup.Groups["Org1"].Policies["TestPolicy"] = &cb.ConfigPolicy{
-		Policy: &cb.Policy{
-			Type: 15,
-		},
-	}
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	err := c.Application().Organization("Org1").RemovePolicy("TestPolicy")
-	gt.Expect(err).To(MatchError("unknown policy type: 15"))
-}
-
-func TestSetApplicationOrgPolicy(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	applicationGroup := newConfigGroup()
-
-	application, _ := baseApplication(t)
-
-	for _, org := range application.Organizations {
-		org.Policies = applicationOrgStandardPolicies()
-
-		orgGroup, err := newOrgConfigGroup(org)
-		gt.Expect(err).NotTo(HaveOccurred())
-
-		applicationGroup.Groups[org.Name] = orgGroup
-	}
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	application.Organizations[0].Policies = applicationOrgStandardPolicies()
-	expectedOrgConfigGroup, _ := newOrgConfigGroup(application.Organizations[0])
-	expectedPolicies := expectedOrgConfigGroup.Policies
-	expectedPolicies["TestPolicy"] = expectedPolicies[EndorsementPolicyKey]
-
-	applicationOrg1 := c.Application().Organization("Org1")
-	err := applicationOrg1.SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "MAJORITY Endorsement"})
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	actualOrg1Policies := applicationOrg1.orgGroup.Policies
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(actualOrg1Policies).To(Equal(expectedPolicies))
-}
-
-func TestSetApplicationOrgPolicyFailures(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	applicationGroup := newConfigGroup()
-
-	application, _ := baseApplication(t)
-	for _, org := range application.Organizations {
-		org.Policies = applicationOrgStandardPolicies()
-
-		orgGroup, err := newOrgConfigGroup(org)
-		gt.Expect(err).NotTo(HaveOccurred())
-
-		applicationGroup.Groups[org.Name] = orgGroup
-	}
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	err := c.Application().Organization("Org1").SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{})
-	gt.Expect(err).To(MatchError("failed to set policy 'TestPolicy': unknown policy type: "))
-}
-
-func TestSetApplicationPolicy(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	application, _ := baseApplication(t)
-
-	applicationGroup, err := newApplicationGroup(application)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		"TestPolicy": {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		},
-	}
-
-	a := c.Application()
-	err = a.SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "MAJORITY Endorsement"})
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedPolicies, err := a.Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
-func TestSetApplicationPolicyFailures(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	application, _ := baseApplication(t)
-
-	applicationGroup, err := newApplicationGroup(application)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	expectedPolicies := application.Policies
-	expectedPolicies["TestPolicy"] = expectedPolicies[EndorsementPolicyKey]
-
-	err = c.Application().SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{})
-	gt.Expect(err).To(MatchError("failed to set policy 'TestPolicy': unknown policy type: "))
-}
-
-func TestRemoveApplicationPolicy(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	application, _ := baseApplication(t)
-
-	applicationGroup, err := newApplicationGroup(application)
-	gt.Expect(err).NotTo(HaveOccurred())
-	applicationGroup.Policies["TestPolicy"] = applicationGroup.Policies[AdminsPolicyKey]
-
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-	}
-
-	a := c.Application()
-	err = a.RemovePolicy("TestPolicy")
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedPolicies, err := a.Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
-func TestRemoveApplicationPolicyFailures(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channelGroup := newConfigGroup()
-	application, _ := baseApplication(t)
-
-	applicationGroup, err := newApplicationGroup(application)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	applicationGroup.Policies[EndorsementPolicyKey] = &cb.ConfigPolicy{
-		Policy: &cb.Policy{
-			Type: 15,
-		},
-	}
-	channelGroup.Groups[ApplicationGroupKey] = applicationGroup
-
-	config := &cb.Config{
-		ChannelGroup: channelGroup,
-	}
-
-	c := New(config)
-
-	err = c.Application().RemovePolicy("TestPolicy")
-	gt.Expect(err).To(MatchError("unknown policy type: 15"))
-}
-
-func TestSetConsortiumOrgPolicy(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	consortiums, _ := baseConsortiums(t)
-
-	consortiumsGroup, err := newConsortiumsGroup(consortiums)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				ConsortiumsGroupKey: consortiumsGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		EndorsementPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		},
-		"TestPolicy": {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		},
-	}
-
-	consortium1Org1 := c.Consortium("Consortium1").Organization("Org1")
-	err = consortium1Org1.SetPolicy("TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "MAJORITY Endorsement"})
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedPolicies, err := consortium1Org1.Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
-func TestSetConsortiumOrgPolicyFailures(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	consortiums, _ := baseConsortiums(t)
-
-	consortiumsGroup, err := newConsortiumsGroup(consortiums)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				ConsortiumsGroupKey: consortiumsGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	for _, test := range []struct {
-		name        string
-		consortium  string
-		org         string
-		policy      Policy
-		expectedErr string
-	}{
-		{
-			name:        "When setting empty policy fails",
-			consortium:  "Consortium1",
-			org:         "Org1",
-			policy:      Policy{},
-			expectedErr: "failed to set policy 'TestPolicy' to consortium org 'Org1': unknown policy type: ",
-		},
-	} {
-		err := c.Consortium(test.consortium).Organization(test.org).SetPolicy("TestPolicy", test.policy)
-		gt.Expect(err).To(MatchError(test.expectedErr))
-	}
-}
-
-func TestRemoveConsortiumOrgPolicy(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	consortiums, _ := baseConsortiums(t)
-
-	consortiums[0].Organizations[0].Policies["TestPolicy"] = Policy{Type: ImplicitMetaPolicyType, Rule: "MAJORITY Endorsement"}
-
-	consortiumsGroup, err := newConsortiumsGroup(consortiums)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				ConsortiumsGroupKey: consortiumsGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		EndorsementPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		},
-	}
-
-	consortium1Org1 := c.Consortium("Consortium1").Organization("Org1")
-	consortium1Org1.RemovePolicy("TestPolicy")
-
-	updatedPolicies, err := consortium1Org1.Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
-func TestSetOrdererPolicy(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	baseOrdererConf, _ := baseSoloOrderer(t)
-
-	ordererGroup, err := newOrdererGroup(baseOrdererConf)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				"Orderer": ordererGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		BlockValidationPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		"TestPolicy": {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Endorsement",
-		},
-	}
-
-	err = c.Orderer().SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "ANY Endorsement"})
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedPolicies, err := c.Orderer().Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
-func TestSetOrdererPolicyFailures(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	baseOrdererConf, _ := baseSoloOrderer(t)
-
-	ordererGroup, err := newOrdererGroup(baseOrdererConf)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				"Orderer": ordererGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	err = c.Orderer().SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{})
-	gt.Expect(err).To(MatchError("failed to set policy 'TestPolicy': unknown policy type: "))
-}
-
-func TestRemoveOrdererPolicy(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	baseOrdererConf, _ := baseSoloOrderer(t)
-	baseOrdererConf.Policies["TestPolicy"] = baseOrdererConf.Policies[AdminsPolicyKey]
-
-	ordererGroup, err := newOrdererGroup(baseOrdererConf)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				"Orderer": ordererGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		BlockValidationPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-	}
-
-	err = c.Orderer().RemovePolicy("TestPolicy")
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedPolicies, err := c.Orderer().Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
-func TestRemoveOrdererPolicyFailures(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	baseOrdererConf, _ := baseSoloOrderer(t)
-	baseOrdererConf.Policies["TestPolicy"] = baseOrdererConf.Policies[AdminsPolicyKey]
-
-	ordererGroup, err := newOrdererGroup(baseOrdererConf)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				OrdererGroupKey: ordererGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	tests := []struct {
-		testName      string
-		ordererGrpMod func(cb.ConfigGroup) *cb.ConfigGroup
-		policyName    string
-		expectedErr   string
-	}{
-		{
-			testName: "when removing blockvalidation policy",
-			ordererGrpMod: func(og cb.ConfigGroup) *cb.ConfigGroup {
-				return &og
-			},
-			policyName:  BlockValidationPolicyKey,
-			expectedErr: "BlockValidation policy must be defined",
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.testName, func(t *testing.T) {
-			gt := NewGomegaWithT(t)
-
-			ordererGroup := tt.ordererGrpMod(*ordererGroup)
-			if ordererGroup == nil {
-				delete(config.ChannelGroup.Groups, OrdererGroupKey)
-			} else {
-				config.ChannelGroup.Groups[OrdererGroupKey] = ordererGroup
-			}
-
-			err = c.Orderer().RemovePolicy(tt.policyName)
-			gt.Expect(err).To(MatchError(tt.expectedErr))
-		})
-	}
-}
-
-func TestSetOrdererOrgPolicy(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	baseOrdererConf, _ := baseSoloOrderer(t)
-
-	ordererGroup, err := newOrdererGroup(baseOrdererConf)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				"Orderer": ordererGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		EndorsementPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		},
-		"TestPolicy": {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Endorsement",
-		},
-	}
-
-	ordererOrg := c.Orderer().Organization("OrdererOrg")
-	err = ordererOrg.SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "ANY Endorsement"})
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedPolicies, err := ordererOrg.Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
-func TestSetOrdererOrgPolicyFailures(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	baseOrdererConf, _ := baseSoloOrderer(t)
-
-	ordererGroup, err := newOrdererGroup(baseOrdererConf)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				"Orderer": ordererGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	err = c.Orderer().Organization("OrdererOrg").SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{})
-	gt.Expect(err).To(MatchError("unknown policy type: "))
-}
-
-func TestRemoveOrdererOrgPolicy(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	baseOrdererConf, _ := baseSoloOrderer(t)
-	baseOrdererConf.Organizations[0].Policies["TestPolicy"] = baseOrdererConf.Organizations[0].Policies[AdminsPolicyKey]
-
-	ordererGroup, err := newOrdererGroup(baseOrdererConf)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				"Orderer": ordererGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		EndorsementPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		},
-	}
-
-	err = c.Orderer().Organization("OrdererOrg").RemovePolicy("TestPolicy")
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedPolicies, err := c.Orderer().Organization("OrdererOrg").Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedPolicies).To(Equal(expectedPolicies))
-}
-
 func TestSetConsortiumChannelCreationPolicy(t *testing.T) {
 	t.Parallel()
 
@@ -845,138 +128,4 @@ func TestSetConsortiumChannelCreationPolicyFailures(t *testing.T) {
 			gt.Expect(err).To(MatchError(tt.expectedErr))
 		})
 	}
-}
-
-func TestSetChannelPolicy(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel, _, err := baseApplicationChannelGroup(t)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: channel,
-	}
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		"TestPolicy": {Type: ImplicitMetaPolicyType, Rule: "ANY Readers"},
-	}
-
-	err = c.Channel().SetPolicy(AdminsPolicyKey, "TestPolicy", Policy{Type: ImplicitMetaPolicyType, Rule: "ANY Readers"})
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedChannelPolicy, err := getPolicies(c.updated.ChannelGroup.Policies)
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedChannelPolicy).To(Equal(expectedPolicies))
-
-	baseChannel := c.original.ChannelGroup
-	gt.Expect(baseChannel.Policies).To(HaveLen(0))
-	gt.Expect(baseChannel.Policies["TestPolicy"]).To(BeNil())
-}
-
-func TestRemoveChannelPolicy(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel, _, err := baseApplicationChannelGroup(t)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: channel,
-	}
-	policies := standardPolicies()
-	err = setPolicies(channel, policies, AdminsPolicyKey)
-	gt.Expect(err).NotTo(HaveOccurred())
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		"Admins": {
-			Type: "ImplicitMeta",
-			Rule: "MAJORITY Admins",
-		},
-		"Writers": {
-			Type: "ImplicitMeta",
-			Rule: "ANY Writers",
-		},
-	}
-
-	err = c.Channel().RemovePolicy(ReadersPolicyKey)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	updatedChannelPolicy, err := c.Channel().Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(updatedChannelPolicy).To(Equal(expectedPolicies))
-
-	originalChannel := c.original.ChannelGroup
-	gt.Expect(originalChannel.Policies).To(HaveLen(3))
-	gt.Expect(originalChannel.Policies[ReadersPolicyKey]).ToNot(BeNil())
-}
-
-func TestRemoveChannelPolicyFailures(t *testing.T) {
-	t.Parallel()
-	gt := NewGomegaWithT(t)
-
-	channel, _, err := baseApplicationChannelGroup(t)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: channel,
-	}
-	policies := standardPolicies()
-	err = setPolicies(channel, policies, AdminsPolicyKey)
-	gt.Expect(err).NotTo(HaveOccurred())
-	channel.Policies[ReadersPolicyKey] = &cb.ConfigPolicy{
-		Policy: &cb.Policy{
-			Type: 15,
-		},
-	}
-	c := New(config)
-
-	err = c.Channel().RemovePolicy(ReadersPolicyKey)
-	gt.Expect(err).To(MatchError("unknown policy type: 15"))
-}
-
-func TestConsortiumOrgPolicies(t *testing.T) {
-	t.Parallel()
-
-	gt := NewGomegaWithT(t)
-
-	consortiums, _ := baseConsortiums(t)
-
-	consortiumsGroup, err := newConsortiumsGroup(consortiums)
-	gt.Expect(err).NotTo(HaveOccurred())
-
-	config := &cb.Config{
-		ChannelGroup: &cb.ConfigGroup{
-			Groups: map[string]*cb.ConfigGroup{
-				ConsortiumsGroupKey: consortiumsGroup,
-			},
-		},
-	}
-
-	c := New(config)
-
-	expectedPolicies := map[string]Policy{
-		ReadersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Readers",
-		},
-		WritersPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "ANY Writers",
-		},
-		AdminsPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Admins",
-		},
-		EndorsementPolicyKey: {
-			Type: ImplicitMetaPolicyType,
-			Rule: "MAJORITY Endorsement",
-		},
-	}
-
-	policies, err := c.Consortium("Consortium1").Organization("Org1").Policies()
-	gt.Expect(err).NotTo(HaveOccurred())
-	gt.Expect(policies).To(Equal(expectedPolicies))
 }
